+++ src/peek.ts
@@ -4,7 +4,7 @@
  * A lightweight JavaScript library that attaches a viewer to a specific node type, allowing you to view the CSS properties, attributes, and size/position.
  * 
  * @file        peek.ts
- * @version     v1.0.0
+ * @version     v1.1.0
  * @author      Bunoon
  * @license     MIT License
  * @copyright   Bunoon 2024
@@ -16,7 +16,8 @@ import { PublicApi } from "./ts/api";
 import { Is } from "./ts/is";
 import { DomElement } from "./ts/dom";
 import { Data } from "./ts/data";
-import { Char, Mode } from "./ts/enum";
+import { Char, IgnoreState, KeyCode, Mode, Value } from "./ts/enum";
+import { Constant } from "./ts/constant";
 
 
 ( () => {
@@ -33,7 +34,6 @@ import { Char, Mode } from "./ts/enum";
     // Variables: Current Process:
     let _current_Process_Options: Options = null!;
     let _current_Process_Elements: HTMLElement[] = [];
-
 
     /*
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
@@ -108,14 +108,14 @@ import { Char, Mode } from "./ts/enum";
         const computedStylesLength: number = computedStyles.length;
 
         for( let styleIndex: number = 0; styleIndex < computedStylesLength; styleIndex++ ) {
-            buildPropertyRow( computedStyles[ styleIndex ], computedStyles.getPropertyValue( computedStyles[ styleIndex ] ) );
+            buildPropertyRow( element, computedStyles[ styleIndex ], computedStyles.getPropertyValue( computedStyles[ styleIndex ] ) );
         }
     }
 
     function buildAttributeProperties( element: HTMLElement ) : void {
         if ( element.hasAttributes() ) {
             for ( let attribute of element.attributes ) {
-                buildPropertyRow( attribute.name, attribute.value );
+                buildPropertyRow( element, attribute.name, attribute.value );
             }
 
         } else {
@@ -126,23 +126,38 @@ import { Char, Mode } from "./ts/enum";
     function buildSizeProperties( element: HTMLElement ) : void {
         const offset: Position = DomElement.getOffset( element );
 
-        buildPropertyRow( "left", offset.left.toString() + "px" );
-        buildPropertyRow( "top", offset.top.toString() + "px" );
-        buildPropertyRow( "width", element.offsetWidth.toString() + "px" );
-        buildPropertyRow( "height", element.offsetHeight.toString() + "px" );
+        buildPropertyRow( element, "left", offset.left.toString() + "px", false );
+        buildPropertyRow( element, "top", offset.top.toString() + "px", false );
+        buildPropertyRow( element, "width", element.offsetWidth.toString() + "px", false );
+        buildPropertyRow( element, "height", element.offsetHeight.toString() + "px", false );
     }
 
-    function buildPropertyRow( propertyNameText: string, propertyValueText: string ) : void {
-        const property: HTMLElement = DomElement.create( _dialog_Contents, "div", "property-row" );
+    function buildPropertyRow( element: HTMLElement, propertyNameText: string, propertyValueText: string, allowEditing: boolean = true ) : void {
+        if ( _current_Process_Options.showOnly!.length === 0 || _current_Process_Options.showOnly!.indexOf( propertyNameText ) > Value.notFound ) {
+            const property: HTMLElement = DomElement.create( _dialog_Contents, "div", "property-row" );
 
-        DomElement.createWithHTML( property, "div", "property-name", propertyNameText );
-        
-        const propertyValue: HTMLElement = DomElement.create( property, "div", "property-value" );
-        const propertyValueInput: HTMLInputElement = DomElement.create( propertyValue, "input" ) as HTMLInputElement;
+            DomElement.createWithHTML( property, "div", "property-name", propertyNameText );
+            
+            const propertyValue: HTMLElement = DomElement.create( property, "div", "property-value" );
+            const propertyValueInput: HTMLInputElement = DomElement.create( propertyValue, "input" ) as HTMLInputElement;
+    
+            propertyValueInput.type = "text";
+            propertyValueInput.value = propertyValueText;
 
-        propertyValueInput.type = "text";
-        propertyValueInput.readOnly = true;
-        propertyValueInput.value = propertyValueText;
+            if ( !_current_Process_Options.allowEditing || !allowEditing ) {
+                propertyValueInput.readOnly = true;
+            } else {
+                propertyValueInput.onkeyup = ( e: KeyboardEvent ) => {
+                    onPropertyValueKeyUp( e, propertyNameText, propertyValueInput, element );
+                }
+            }
+        }
+    }
+
+    function onPropertyValueKeyUp( e: KeyboardEvent, propertyName: string, input: HTMLInputElement, element: HTMLElement ) {
+        if ( e.code === KeyCode.enter ) {
+            element.style.setProperty( propertyName, input.value );
+        }
     }
 
 
@@ -170,11 +185,15 @@ import { Char, Mode } from "./ts/enum";
     }
 
     function buildNodeEvent( element: HTMLElement ) : void {
-        element.addEventListener( "mousemove", ( e ) => {
-            onNodeMouseOver( e, element );
-        } );
+        const attributeValue: string = element.getAttribute( Constant.PEEK_JS_IGNORE_STATE_ATTRIBUTE )!;
 
-        _current_Process_Elements.push( element );
+        if ( !Is.definedString( attributeValue ) ?? attributeValue !== IgnoreState.ignore ) {
+            element.addEventListener( "mousemove", ( e ) => {
+                onNodeMouseOver( e, element );
+            } );
+    
+            _current_Process_Elements.push( element );
+        }
     }
 
     function removeNodeEvents() : void {
@@ -222,6 +241,8 @@ import { Char, Mode } from "./ts/enum";
         options.nodeType = Data.getDefaultStringOrArray( options.nodeType, [] );
         options.mode = Data.getDefaultNumber( options.mode, Mode.css );
         options.titleText = Data.getDefaultString( options.titleText, Char.empty );
+        options.showOnly = Data.getDefaultStringOrArray( options.showOnly, [] );
+        options.allowEditing = Data.getDefaultBoolean( options.allowEditing, false );
 
         return options;
     }
@@ -323,7 +344,7 @@ import { Char, Mode } from "./ts/enum";
          */
 
         getVersion: function (): string {
-            return "1.0.0";
+            return "1.1.0";
         }
     };