+++ src/ts/enum.ts
@@ -4,7 +4,7 @@
  * A lightweight JavaScript library that attaches a viewer to a specific node type, allowing you to view the CSS properties, attributes, and size/position.
  * 
  * @file        enum.ts
- * @version     v1.0.0
+ * @version     v1.1.0
  * @author      Bunoon
  * @license     MIT License
  * @copyright   Bunoon 2024
@@ -20,4 +20,16 @@ export const enum Mode {
     css = 1,
     attributes = 2,
     size = 3,
+}
+
+export const enum Value {
+    notFound = -1,
+}
+
+export const enum KeyCode {
+    enter = "Enter",
+}
+
+export const enum IgnoreState {
+    ignore = "ignore",
 }