+++ docs/CHANGE_LOG.md
@@ -2,6 +2,7 @@
 
 ## Version 1.1.0:
 - Added new binding option called "showOnly" (excepts a string, or string array), which states the properties that should be shown in the dialog.
+- Added new binding option called "allowEditing" (defaults to false), which allows CSS and Attribute values to be entered (simply type and press Enter).
 
 <br>