+++ docs/CHANGE_LOG.md
@@ -1,4 +1,12 @@
 # Peek.js - Change Log:
 
+## Version 1.1.0:
+- Added a new binding option called "showOnly" (excepts a string, or string array), which states the properties that should be shown in the dialog.
+- Added a new binding option called "allowEditing" (defaults to false), which allows CSS and Attribute values to be entered (simply type and press Enter).
+- Fixed a fault that caused the dialog elements to be included when viewing specific node types.
+- Added a new data attribute called "data-peek-js-ignore-state", which can be set to "ignore" on elements so that they cannot be used by the viewer.
+
+<br>
+
 ## Version 1.0.0:
 - Everything :)